refactor: merge dev/1.0
