+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@oasis-engine/core",
-  "version": "0.9.0-beta.66",
+  "version": "1.0.0-alpha.0",
   "publishConfig": {
     "access": "public",
     "registry": "https://registry.npmjs.org"
