+++ packages/core/src/Engine.ts
@@ -102,6 +102,8 @@ export class Engine extends EventDispatcher {
   _spriteMaskManager: SpriteMaskManager;
   /** @internal */
   _canSpriteBatch: boolean = true;
+  /** @internal */
+  _fontMap: Record<string, Font> = {};
   /** @internal @todo: temporary solution */
   _macroCollection: ShaderMacroCollection = new ShaderMacroCollection();
 
@@ -363,7 +365,8 @@ export class Engine extends EventDispatcher {
     this._resourceManager._destroy();
     this._magentaTexture2D.destroy(true);
     this._magentaTextureCube.destroy(true);
-    this._textDefaultFont.destroy(true);
+    this._textDefaultFont = null;
+    this._fontMap = null;
 
     this.inputManager._destroy();
     this.dispatch("shutdown", this);
