+++ packages/loader/src/gltf/extensions/index.ts
@@ -6,9 +6,12 @@ import "./KHR_materials_pbrSpecularGlossiness";
 import "./KHR_materials_sheen";
 import "./KHR_materials_transmission";
 import "./KHR_materials_unlit";
-import "./KHR_materials_variants";
+export type { IGLTFExtensionVariants } from "./KHR_materials_variants";
 import "./KHR_materials_volume";
 import "./KHR_mesh_quantization";
 import "./KHR_texture_basisu";
 import "./KHR_texture_transform";
 import "./OASIS_materials_remap";
+
+export { GLTFExtensionParser, GLTFExtensionMode } from "./GLTFExtensionParser";
+export * from "./GLTFExtensionSchema";
