+++ packages/loader/src/gltf/index.ts
@@ -1,0 +1,6 @@
+export { GLTFPipeline } from "./GLTFPipeline";
+export { GLTFResource } from "./GLTFResource";
+export { GLTFUtil } from "./GLTFUtil";
+export * from "./parser";
+export * from "./extensions/index";
+export type { IMaterial, IMeshPrimitive, ITextureInfo, INode, GLTFExtensionOwnerSchema } from "./GLTFSchema";
