+++ packages/loader/src/gltf/parser/GLTFAnimationParser.ts
@@ -14,12 +14,12 @@ import {
 } from "@oasis-engine/core";
 import { Quaternion, Vector3, Vector4 } from "@oasis-engine/math";
 import { GLTFUtil } from "../GLTFUtil";
-import { AccessorType, AnimationChannelTargetPath, AnimationSamplerInterpolation, IAnimationChannel } from "../Schema";
-import { Parser } from "./Parser";
-import { ParserContext } from "./ParserContext";
+import { AccessorType, AnimationChannelTargetPath, AnimationSamplerInterpolation, IAnimationChannel } from "../GLTFSchema";
+import { GLTFParser } from "./GLTFParser";
+import { GLTFParserContext } from "./GLTFParserContext";
 
-export class AnimationParser extends Parser {
-  parse(context: ParserContext): AssetPromise<AnimationClip[]> {
+export class GLTFAnimationParser extends GLTFParser {
+  parse(context: GLTFParserContext): AssetPromise<AnimationClip[]> {
     const { glTF, buffers, glTFResource } = context;
     const { entities } = glTFResource;
     const { animations, accessors, bufferViews } = glTF;
